Created 'PlaybackComponent'.
